Add support for basic watch mode
Resolves #14.

Adds npm-watch for re-running `eslint` and `tsc`. ESLint does not have a
watch mode built in, and TypeScript hard-flushes the buffer every time
it re-runs the type-check. This makes builds noticeably slower; we may
at some point in the future (post-MVP) want to extract that into a local
type-checking script which behaves like TS's built-in watch mode, but
with better stdio management (cf. ember-cli-typescript, which does just
that, albeit for other reasons).